CI: Enforce -Wall and -Werror
Part 2 of: https://github.com/ornladios/ADIOS2/issues/2766

Followup of https://github.com/ornladios/ADIOS2/pull/2639/files

Notes from this PR:

Wall -Werror enabled for C++ in all builds but:

1. Accent + XL builds
2. Power8 builds
3. Debian buids
4. Using intel compiler translation units that include `nlohmann_json`

+++ source/adios2/toolkit/format/bp5/BP5Base.cpp
@@ -25,7 +25,7 @@ namespace format
 
 void BP5Base::FFSBitfieldSet(struct FFSMetadataInfoStruct *MBase, int Bit)
 {
-    int Element = Bit / (sizeof(size_t) * 8);
+    size_t Element = Bit / (sizeof(size_t) * 8);
     int ElementBit = Bit % (sizeof(size_t) * 8);
     if (static_cast<size_t>(Element) >= MBase->BitFieldCount)
     {
@@ -40,7 +40,7 @@ void BP5Base::FFSBitfieldSet(struct FFSMetadataInfoStruct *MBase, int Bit)
 
 int BP5Base::FFSBitfieldTest(struct FFSMetadataInfoStruct *MBase, int Bit)
 {
-    int Element = Bit / (sizeof(size_t) * 8);
+    size_t Element = Bit / (sizeof(size_t) * 8);
     int ElementBit = Bit % (sizeof(size_t) * 8);
     if (static_cast<size_t>(Element) >= MBase->BitFieldCount)
     {
